fix(invoice): clear return-link fields on type change, enrich mutation responses
Auto-clear settled_via, applied_to_invoice_id, and refunds_invoice_id when an
update moves an invoice's type away from 'return', mirroring the existing
service_month auto-clear pattern; otherwise the link fields strand and keep
consuming the target's applied-amount cap. Type changes back to 'return' no
longer resurrect a previously cleared link.

Create/update invoice responses now enrich paid_with_returns the same way
list/GET already do, so the frontend detail view doesn't go stale on save.

applied_return_summaries amounts now pass through the money() quantizer to
avoid sub-cent float artifacts in paid_with_returns JSON.

Document the accepted read-then-write race on the applied-amount cap check
(same class as the pre-existing refunds cap check) with no locking change.

+++ app/infrastructure/adapters/sqlalchemy_invoice.py
@@ -8,6 +8,7 @@ from uuid import UUID
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
+from app.application.invoice.dtos import money
 from app.application.invoice.ports import IInvoiceRepository
 from app.domain.entities.invoice import Invoice, InvoiceType, RefundableStatus
 from app.domain.exceptions.invoice_exceptions import (
@@ -506,7 +507,7 @@ class SQLAlchemyInvoiceRepository(IInvoiceRepository):
         result: dict[UUID, list[dict]] = {}
         for target_id, invoice_number, items in rows:
             result.setdefault(target_id, []).append(
-                {"invoice_number": invoice_number, "total_amount": float(_items_total(items))}
+                {"invoice_number": invoice_number, "total_amount": money(_items_total(items))}
             )
         return result
